Refactor data loading out of app.py
